refactor tests to be in respective __tests__ directory
