+++ _rules/__tests__/frontmatter.js
@@ -1,5 +1,5 @@
 
-const describeRule = require('../test-utils/describe-rule')
+const describeRule = require('../../test-utils/describe-rule')
 const { contributors } = require('./../../package.json')
 const contributorsNames = contributors.map(contributor => contributor.name.toLowerCase())
